Move `LiveHackrNewsStoreSpy` into a helpers shared-level file
